pyproject.toml
[project]
name = "yaai"
version = "0.1.0"
description = "Yet Another AI monitoring — SDK and self-hosted ML model monitoring platform"
readme = "README.md"
license = "LicenseRef-Elastic-2.0"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
]

[project.optional-dependencies]
gcp = [
    "google-auth>=2.0",
]
server = [
    "alembic>=1.18.3",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "authlib>=1.3.0",
    "bcrypt>=4.2.0",
    "cachetools>=5.5.0",
    "fastapi>=0.128.4",
    "google-auth>=2.0",
    "itsdangerous>=2.2.0",
    "numpy>=2.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.9.0",
    "requests>=2.20.0",
    "scikit-learn>=1.6.0",
    "scipy>=1.17.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]

[project.urls]
Repository = "https://github.com/Maxl94/yaai"
Documentation = "https://github.com/Maxl94/yaai/tree/main/docs"
Issues = "https://github.com/Maxl94/yaai/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["yaai"]

[tool.ruff]
line-length = 120

lint.select = [
    "E",  # Pyflakes (Error detection)
    "F",  # Pycodestyle (Formatting errors)
    "B",  # Bugbear (Best practices & potential issues)
    "C",  # Complexity (Cyclomatic complexity)
    "I",  # isort (Import sorting)
    "PL", # Pylint (Code quality checks)
    "UP", # pyupgrade (Modern Python usage)
    "S",  # Bandit (Security issues)
    "TID",    # Type hinting checks (flake8-type-checking)
    "C4",     # Complexity analysis (McCabe)
    "RUF100", # "noqa" rule ignore validation
]

# Ignore specific rules that may be too strict
lint.ignore = [
    "E501",    # Ignore long lines (handled by formatter)
    "PLR0913", # Too many function arguments (sometimes necessary)
    "B008",    # Function calls in argument defaults (can be intentional)
    "D100",    # Missing docstring in public module
    "D104",    # Missing docstring in public package
    "S101",    # Use of assert statement (can be intentional)
]


exclude = [
    ".git",
    ".ruff_cache",
    ".venv",
    "__pypackages__",
    "notebooks/databricks",
    "data",
    "docs",
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["yaai"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S108", "PLR2004", "PLC0415"]
"yaai/client.py" = ["PLR2004"]
"yaai/server/drift/*" = ["PLR2004"]
"yaai/server/scheduler.py" = ["PLC0415"]
"yaai/server/schemas/job.py" = ["PLC0415"]
"yaai/server/auth/*" = ["S105", "S106", "PLW0603", "PLC0415", "PLR2004", "PLR0911"]
"yaai/server/main.py" = ["S105", "S106"]
"scripts/*" = ["PLR0915", "PLR2004", "C901"]

[tool.pyscan.complexity]
max_complexity = 15

[tool.pyscan.dead_code]
min_severity = "warning"

[tool.pyscan.output]
directory = "reports"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[dependency-groups]
dev = [
    "yaai[server]",
    "aiosqlite>=0.22.1",
    "greenlet>=3.3.1",
    "ipython>=9.10.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "pyscn>=1.10.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.6.0",
    "mkdocstrings[python]>=0.27.0",
    "mkdocs-render-swagger-plugin>=0.1.2",
    "testcontainers[postgres]>=4.0.0",
]
